chore: refactor getGasReceiverContractAddress by removing duplicated functions and move to AxelarQueryAPI
